Update release pipeline and change-release-version script

+++ scripts/change-release-version.sh
@@ -1,10 +1,19 @@
-node change-release-version.js
+#!/bin/bash
+
+# Check if --version is provided
+if [[ $# -eq 0 || "$1" != "--version" ]]; then
+    echo "❌ Error: --version argument is required."
+    exit 1
+fi
+
+npm install
+node change-release-version.js "$@"
 cd ../documentation
 rm -rf ./node-modules
 npm install
 npm run build
 cd ..
-rm -rf ../docs
+rm -rf ./docs
 mv ./documentation/build ./docs 
 git add ./documentation
 git add ./docs
